+++ src/w5100.cpp
@@ -278,14 +278,7 @@ uint16_t EtherSia_W5100::readFrame(uint8_t *buffer, uint16_t bufsize)
         wizchip_recv_data(buffer, data_len);
         setSn_CR(Sn_CR_RECV);
 
-        // W5100 doesn't have any built-in MAC address filtering
-        if ((buffer[0] & 0x01) || memcmp(&buffer[0], _localMac, 6) == 0)
-        {
-            // Addressed to an Ethernet multicast address or our unicast address
-            return data_len;
-        } else {
-            return 0;
-        }
+        return data_len;
     }
 
     return 0;