+++ tests/libarduino/Progmem.h
@@ -10,7 +10,8 @@ class __FlashStringHelper;
 #define memcpy_P(dst, src, n) memcpy(dst, src, n)
 #define memcmp_P(p1, p2, n) memcmp(p1, p2, n)
 #define strcmp_P(s1, s2) strcmp(s1, s2)
-#define strcpy_P(s1, s2) strcpy(s1, s2)
+#define strcpy_P(dst, src) strcpy(dst, src)
+#define strncpy_P(dst, src, len) strncpy(dst, src, len)
 #define strlen_P(str) strlen(str)
 #define pgm_read_byte(addr) *(addr)